add kdtree

+++ src/lanecluster.cpp
@@ -8,7 +8,7 @@ LaneCluster::LaneCluster()
 
 void LaneCluster::apply_lane_feats_cluster(const cv::Mat &binary_seg_result, const cv::Mat &instance_seg_result,std::vector<inner_type::Lane> &lane_coords,cv::Mat &mask)
 {
-    std::vector<std::vector<float>> lane_embedding_feats;
+    std::vector<std::vector<double>> lane_embedding_feats;
     std::vector<inner_type::LanePoint> coord;
     std::vector<int> db_labels;
     std::vector<int> unique_labels;
@@ -56,7 +56,7 @@ void LaneCluster::apply_lane_feats_cluster(const cv::Mat &binary_seg_result, con
     int bbb=2;
 }
 
-void LaneCluster::_embedding_feats_dbscan_cluster(const std::vector<std::vector<float>> &lane_embedding_feats,std::vector<int> &db_labels,std::vector<int> &unique_labels)
+void LaneCluster::_embedding_feats_dbscan_cluster(const std::vector<std::vector<double>> &lane_embedding_feats,std::vector<int> &db_labels,std::vector<int> &unique_labels)
 {
     sp_dbscan->cluster(lane_embedding_feats);
     db_labels=sp_dbscan->labels_;
@@ -64,7 +64,7 @@ void LaneCluster::_embedding_feats_dbscan_cluster(const std::vector<std::vector<
 }
 
 
-void LaneCluster::_get_lane_embedding_feats(const cv::Mat &binary_seg_result, const cv::Mat &instance_seg_result,std::vector<std::vector<float>> &lane_embedding_feats,std::vector<inner_type::LanePoint> &lane_coordinates)
+void LaneCluster::_get_lane_embedding_feats(const cv::Mat &binary_seg_result, const cv::Mat &instance_seg_result,std::vector<std::vector<double>> &lane_embedding_feats,std::vector<inner_type::LanePoint> &lane_coordinates)
 {
     for (size_t nrow = 0; nrow < binary_seg_result.rows; nrow++)
     {
@@ -78,7 +78,7 @@ void LaneCluster::_get_lane_embedding_feats(const cv::Mat &binary_seg_result, co
                 float r=instance_ptr[ncol][0];
                 float g=instance_ptr[ncol][1];
                 float b=instance_ptr[ncol][2];
-                std::vector<float> feat;
+                std::vector<double> feat;
                 feat.push_back(r);
                 feat.push_back(g);
                 feat.push_back(b);
